task BC-3448

+++ src/main/java/school/faang/user_service/controller/SubscriptionController.java
@@ -32,7 +32,7 @@ public class SubscriptionController {
         return subscriptionService.getFollowers(followeeId, filter);
     }
 
-    public int getFollowersCount(long followerId){
+    public long getFollowersCount(long followerId){
         return subscriptionService.getFollowersCount(followerId);
     }
 }
